Refactor the error message system
