+++ paddle/cinn/ir/ir_schedule.h
@@ -24,24 +24,11 @@
 #include "paddle/cinn/ir/ir_mutator.h"
 #include "paddle/cinn/ir/schedule_desc.h"
 #include "paddle/cinn/ir/tensor.h"
+#include "paddle/cinn/utils/error.h"
 #include "paddle/cinn/utils/random_engine.h"
 
 namespace cinn {
 namespace ir {
-
-/**
- *  \brief Indicates the level of printing error message in the current Schedule
- */
-enum class ScheduleErrorMessageLevel : int32_t {
-  /** \brief  Print an error message in short mode.
-   * Short mode shows which and where the schedule error happens*/
-  kGeneral = 0,
-  /** \brief Print an error message in detailed mode.
-   * Detailed mode shows which and where the schedule error happens, and the
-   * schedule input parameters.
-   */
-  kDetailed = 1,
-};
 
 /**
  * A struct representing a module that contains Expr. This struct is only used
@@ -85,8 +72,8 @@ class IRSchedule {
   explicit IRSchedule(const ModuleExpr& modexpr,
                       utils::LinearRandomEngine::StateType rand_seed = -1,
                       bool debug_flag = false,
-                      ScheduleErrorMessageLevel err_msg_level =
-                          ScheduleErrorMessageLevel::kGeneral);
+                      utils::ErrorMessageLevel err_msg_level =
+                          utils::ErrorMessageLevel::kGeneral);
   IRSchedule(ir::ModuleExpr&& mod_expr,
              ScheduleDesc&& trace,
              utils::LinearRandomEngine::StateType rand_seed = -1);
